backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
openai>=1.82.0
langchain==0.1.0
langchain-community==0.0.16
langchain-core==0.1.18
langchain-openai==0.0.5
pydantic==2.5.3
aiohttp==3.9.1
python-multipart==0.0.6
certifi==2024.2.2
faiss-cpu==1.7.4